+++ example/tech/initTexelCopy.ts
@@ -137,7 +137,7 @@ fn fs_main(in: VertexOutput) -> @location(0) vec4f {
 
     let dispatch1: RenderProperty;
     {
-        const indexData = new Int16Array([0, 1, 2, 0, 2, 3]);
+        const indexData = new Uint32Array([0, 1, 2, 0, 2, 3]);
         const indexBuffer = compiler.createIndexBuffer({
             rawData: indexData
         });
